crewai/orchestrator/pyproject.toml
[project]
name = "orchestrator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Muhammad Fasih", email = "programmer.fasih.18@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "crewai>=0.102.0",
    "litellm>=1.60.2",
    "python-dotenv>=1.0.1",
]

[project.scripts]
orchestrator = "orchestrator:main"
own = "orchestrator.own:main"
llm = "orchestrator.llm:main"
repo = "orchestrator.repo:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
